플레이어 점프 수정

+++ Alto_Adventure/Obj.cpp
@@ -35,8 +35,8 @@ void CObj::Fall()
 
 void CObj::Update_Rect()
 {
-	m_tRect.left = m_tInfo.vPos.x - m_tInfo.vSize.x / 2.f;
-	m_tRect.right = m_tInfo.vPos.x + m_tInfo.vSize.x / 2.f;
-	m_tRect.top = m_tInfo.vPos.y - m_tInfo.vSize.y / 2.f;
-	m_tRect.bottom = m_tInfo.vPos.y + m_tInfo.vSize.y / 2.f;
+	m_tRect.left = LONG(m_tInfo.vPos.x - m_tInfo.vSize.x / 2.f);
+	m_tRect.right = LONG(m_tInfo.vPos.x + m_tInfo.vSize.x / 2.f);
+	m_tRect.top = LONG(m_tInfo.vPos.y - m_tInfo.vSize.y / 2.f);
+	m_tRect.bottom = LONG(m_tInfo.vPos.y + m_tInfo.vSize.y / 2.f);
 }
